Feature: The Form-Generator now uses Topic-Descriptions as a hint for label/input field pair if a "Property"-Topic is related to a "Topic" via the new AssociationType "Helptext".
Identification:
When "Property"-Topics contain a quotation mark in their name, the property is not saved to database. Now an error warning is logged.

IMPORTANT:
This commit contains a patch which updates the CM from 2.19 to 2.20.
Apply this patch if you want to update DeepaMehta from 2.0b8 to DeepaMehta rev387 while keeping your content